backend/requirements.txt
# backend/requirements.txt

fastapi==0.89.0
sqlalchemy[asyncio]==2.0.0
asyncpg==0.25.0
pydantic==2.0.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
uvicorn==0.19.0